Move UserLocalDataSource.kt to the new package - store.
