+++ app/src/main/java/com/appunite/loudius/di/DataSourceModule.kt
@@ -1,7 +1,7 @@
 package com.appunite.loudius.di
 
 import android.content.Context
-import com.appunite.loudius.domain.UserLocalDataSource
+import com.appunite.loudius.domain.store.UserLocalDataSource
 import com.appunite.loudius.network.datasource.AuthDataSource
 import com.appunite.loudius.network.datasource.AuthNetworkDataSource
 import com.appunite.loudius.network.datasource.PullRequestDataSource
