+++ app/src/main/java/com/appunite/loudius/domain/store/UserLocalDataSource.kt
@@ -1,4 +1,4 @@
-package com.appunite.loudius.domain
+package com.appunite.loudius.domain.store
 
 import android.content.Context
 import android.content.SharedPreferences
